Admin login
You can now login as an admin and if you do you will be able to see and click the stock page tab in de nav menu.

Changed stock page css so it's a bit better when viewed on a smaller screen.

Also changed a little css for the nav menu (with how much of the menu I removed this is kind of redundant)