Profile page
Made something of a profile page. (main focus)
Mostly when you change things in the profile page while being logged in it'll effectivly update the database.
There should probably be some limiting factors for this but I just want to lay this beast to rest for now.
Some css because of course.
(I have once more become indecisive on how I'd organise my css)
Changed some things about the login button in the nav menu.

I really want to make it so that the process of placing/updating your order/profile page info happens via the process page because there are some similarities between the two, but I'm unsure how to best work this out.