Process, profile and formatting
Primarily wanted to push this one out because I used a formatter that changed a lot of spacing and it kinda makes it difficult to tell what else I did.

Other then that the profile page got some minimal changes but the biggest one is the products page, altough it's not finished.